[tests] Moved tests to dedicated file for consistency and clarity
Also removed the override_settings call which wasn't needed.
